requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
google-genai>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
